v2025.8.1

+++ software/python-package/shepherd_sheep/shared_mem_iv_input.py
@@ -195,8 +195,8 @@ class SharedMemIVInput:
         # transform raw ADC data to SI-Units -> the virtual-source-emulator in PRU expects uV and nV
         if cal:
             # option to disable scaling here if already done (performance improvement)
-            data.voltage = cal.voltage.raw_to_si(data.voltage).astype("u4")
-            data.current = cal.current.raw_to_si(data.current).astype("u4")
+            data.voltage = cal.voltage.raw_to_si(data.voltage).astype(np.uint32)
+            data.current = cal.current.raw_to_si(data.current).astype(np.uint32)
         # interweave data (voltage | current in parallel)
         iv_data = np.empty((2 * len(data),), dtype=data.voltage.dtype)
         iv_data[0::2] = data.voltage[: len(data)]
